Move class TvShow to own file
